Added page authentication and some various tweaks and fixes.
