+++ handcrafted-haven/src/app/products/[productId]/edit/page.tsx
@@ -3,6 +3,8 @@ import {fetchProductById} from '@/app/lib/data';
 import { notFound } from 'next/navigation';
 import Navbar from '@/components/Navbar';
 import Footer from '@/components/Footer';
+import { getUserData } from "@/app/lib/actions";
+import NotAuthorized from "@/components/NotAuthorized";
 
 export default async function Page( props: {params: Promise<{ productId: string }> }) {
     const params = await props.params;
@@ -12,6 +14,15 @@ export default async function Page( props: {params: Promise<{ productId: string 
     if (!product) {
         notFound();
     }
+
+    const user = await getUserData();
+        if (!user || user.user_type === "basic" || user.id !== product.user_id) {
+            return (
+                <NotAuthorized/>
+            )
+        };
+
+    
 
     return (
         <div className="min-h-screen flex flex-col justify-between">
