+++ handcrafted-haven/src/app/products/[productId]/review/[reviewId]/edit/page.tsx
@@ -3,6 +3,8 @@ import {fetchReviewById} from '@/app/lib/data';
 import { notFound } from 'next/navigation';
 import Navbar from '@/components/Navbar';
 import Footer from '@/components/Footer';
+import { getUserData } from "@/app/lib/actions";
+import NotAuthorized from "@/components/NotAuthorized";
 
 export default async function Page( props: {params: Promise<{ reviewId: string }> }) {
     const params = await props.params;
@@ -14,6 +16,13 @@ export default async function Page( props: {params: Promise<{ reviewId: string }
     if (!review) {
         notFound();
     }
+
+    const user = await getUserData();
+    if (!user || user.id !== review.user_id) {
+        return (
+            <NotAuthorized/>
+        )
+    };
 
     return (
         <div className="min-h-screen flex flex-col justify-between">
