+++ handcrafted-haven/src/app/products/[productId]/review/create/page.tsx
@@ -3,12 +3,21 @@ import {fetchProductById} from '@/app/lib/data';
 import { notFound } from 'next/navigation';
 import Navbar from '@/components/Navbar';
 import Footer from '@/components/Footer';
+import { getUserData } from "@/app/lib/actions";
+import NotAuthorized from "@/components/NotAuthorized";
 
 export default async function Page( props: {params: Promise<{ productId: string }> }) {
     const params = await props.params;
     const id = params.productId;
     console.log(id);
     const product= await fetchProductById(id);
+
+    const user = await getUserData();
+        if (!user) {
+            return (
+                <NotAuthorized/>
+            )
+        };
 
     if (!product) {
         notFound();
